refactor: useProducts custom hook 생성
